requirements.txt
grpcio
grpcio-tools
torch
torchvision
Pillow==9.3.0